add interval to settings, refactoring

+++ include/settings.h
@@ -6,13 +6,13 @@ void settingsProcess();
 float getSettingsTemperature();
 float getSettingsDelta();
 bool getDeviceState();
-long unsigned int getSensorsCheckInterval();
+uint32_t getSensorsCheckInterval();
 float getTemperatureLowBound();
 float getTemperatureHighBound();
 
 void setSettingsTemperature(float value);
 void setSettingsDelta(float value);
-void setSensorsCheckInterval(long unsigned int value);
+void setSensorsCheckInterval(uint32_t value);
 
 void toggleDeviceState();
 void toggleDeviceState(bool state);
